Refactored setting classes into correct namespace.
